gitserver: Revert hexagonal architecture
We started making this move at some point, but we don't anticipate to follow this pattern at this time - so reverting the existing start of it to streamline our code more.
We can rethink this pattern when we have time.
