+++ cmd/frontend/backend/repos.go
@@ -297,7 +297,7 @@ func (s *repos) GetCommit(ctx context.Context, repo *types.Repo, commitID api.Co
 
 	s.logger.Debug("GetCommit", log.String("repo", string(repo.Name)), log.String("commitID", string(commitID)))
 
-	if !gitserver.IsAbsoluteRevision(string(commitID)) {
+	if !gitdomain.IsAbsoluteRevision(string(commitID)) {
 		return nil, errors.Errorf("non-absolute CommitID for Repos.GetCommit: %v", commitID)
 	}
