+++ cmd/frontend/graphqlbackend/git_object.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"sync"
 
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
@@ -42,7 +41,7 @@ func (GitObjectID) ImplementsGraphQLType(name string) bool {
 }
 
 func (id *GitObjectID) UnmarshalGraphQL(input any) error {
-	if input, ok := input.(string); ok && gitserver.IsAbsoluteRevision(input) {
+	if input, ok := input.(string); ok && gitdomain.IsAbsoluteRevision(input) {
 		*id = GitObjectID(input)
 		return nil
 	}
