+++ cmd/gitserver/internal/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
     srcs = [
         "cleanup.go",
         "clone.go",
-        "commands.go",
         "disk.go",
         "gitservice.go",
         "list_gitolite.go",
@@ -48,7 +47,6 @@ go_library(
         "//internal/featureflag",
         "//internal/fileutil",
         "//internal/gitserver",
-        "//internal/gitserver/adapters",
         "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
         "//internal/gitserver/search",
@@ -122,6 +120,7 @@ go_test(
         "//internal/database/dbtest",
         "//internal/extsvc/gitolite",
         "//internal/gitserver",
+        "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
         "//internal/gitserver/v1:gitserver",
         "//internal/grpc",
