+++ cmd/gitserver/internal/server_test.go
@@ -38,6 +38,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database/dbmocks"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
 	proto "github.com/sourcegraph/sourcegraph/internal/gitserver/v1"
 	"github.com/sourcegraph/sourcegraph/internal/grpc"
@@ -1657,5 +1658,115 @@ error: Could not read d24d09b8bc5d1ea2c3aa24455f4578db6aa3afda`,
 		if stdErrIndicatesCorruption(stderr) {
 			t.Errorf("should not contain corrupt line:\n%s", stderr)
 		}
+	}
+}
+
+func TestGetObject(t *testing.T) {
+	sampleSHA := "a03384f3a47acae11478ba7b4a6f331564938d4f"
+	sampleOID, err := decodeOID(sampleSHA)
+	if err != nil {
+		t.Fatal(err)
+	}
+	repoName := api.RepoName("github.com/sourcegraph/sourcegraph")
+
+	tests := []struct {
+		name string
+
+		revParse      revParseFunc
+		getObjectType getObjectTypeFunc
+
+		repo       api.RepoName
+		objectName string
+		wantObject *gitdomain.GitObject
+		wantError  error
+	}{
+		{
+			name: "Happy path",
+			revParse: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error) {
+				return sampleSHA, nil
+			},
+			getObjectType: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error) {
+				return gitdomain.ObjectTypeCommit, nil
+			},
+			repo:       repoName,
+			objectName: "abc",
+			wantObject: &gitdomain.GitObject{ID: sampleOID, Type: gitdomain.ObjectTypeCommit},
+			wantError:  nil,
+		},
+		{
+			name: "Revparse repo doesn't exist",
+			revParse: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error) {
+				return "", &gitdomain.RepoNotExistError{}
+			},
+			getObjectType: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error) {
+				return gitdomain.ObjectTypeCommit, nil
+			},
+			repo:       repoName,
+			objectName: "abc",
+			wantObject: nil,
+			wantError:  &gitdomain.RepoNotExistError{},
+		},
+		{
+			name: "Unknown revision",
+			revParse: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error) {
+				return "unknown revision: foo", errors.New("unknown revision")
+			},
+			getObjectType: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error) {
+				return gitdomain.ObjectTypeCommit, nil
+			},
+			repo:       repoName,
+			objectName: "abc",
+			wantObject: nil,
+			wantError: &gitdomain.RevisionNotFoundError{
+				Repo: repoName,
+				Spec: "abc",
+			},
+		},
+		{
+			name: "HEAD treated as revision not found",
+			revParse: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error) {
+				return "HEAD", nil
+			},
+			getObjectType: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error) {
+				return gitdomain.ObjectTypeCommit, nil
+			},
+			repo:       repoName,
+			objectName: "abc",
+			wantObject: nil,
+			wantError: &gitdomain.RevisionNotFoundError{
+				Repo: repoName,
+				Spec: "abc",
+			},
+		},
+		{
+			name: "Bad commit",
+			revParse: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, rev string) (string, error) {
+				return "not_valid_commit", nil
+			},
+			getObjectType: func(ctx context.Context, rcf *wrexec.RecordingCommandFactory, repo api.RepoName, dir common.GitDir, objectID string) (gitdomain.ObjectType, error) {
+				return gitdomain.ObjectTypeCommit, nil
+			},
+			repo:       repoName,
+			objectName: "abc",
+			wantObject: nil,
+			wantError: &gitdomain.BadCommitError{
+				Repo:   repoName,
+				Spec:   "abc",
+				Commit: api.CommitID("not_valid_commit"),
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			ctx := context.Background()
+			obj, err := getObject(ctx, wrexec.NewNoOpRecordingCommandFactory(), t.TempDir(), tc.getObjectType, tc.revParse, tc.repo, tc.objectName)
+			if diff := cmp.Diff(tc.wantObject, obj); diff != "" {
+				t.Errorf("Object does not match: %v", diff)
+			}
+			if diff := cmp.Diff(tc.wantError, err); diff != "" {
+				t.Errorf("Error does not match: %v", diff)
+			}
+		})
 	}
 }
