+++ internal/codeintel/resolvers/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
     visibility = ["//:__subpackages__"],
     deps = [
         "//internal/api",
-        "//internal/gitserver",
+        "//internal/gitserver/gitdomain",
         "//internal/gqlutil",
         "//internal/markdown",
         "//internal/types",
