+++ internal/codeintel/resolvers/git.go
@@ -6,7 +6,7 @@ import (
 	"github.com/graph-gophers/graphql-go"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -40,7 +40,7 @@ func (GitObjectID) ImplementsGraphQLType(name string) bool {
 }
 
 func (id *GitObjectID) UnmarshalGraphQL(input any) error {
-	if input, ok := input.(string); ok && gitserver.IsAbsoluteRevision(input) {
+	if input, ok := input.(string); ok && gitdomain.IsAbsoluteRevision(input) {
 		*id = GitObjectID(input)
 		return nil
 	}
