--- internal/gitserver/adapters/BUILD.bazel
@@ -1,16 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "adapters",
-    srcs = ["git.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/gitserver/adapters",
-    visibility = ["//:__subpackages__"],
-    deps = [
-        "//internal/api",
-        "//internal/gitserver/gitdomain",
-        "//internal/gitserver/protocol",
-        "//internal/wrexec",
-        "//lib/errors",
-        "@com_github_sourcegraph_log//:log",
-    ],
-)
