+++ internal/gitserver/gitdomain/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "errors.go",
         "exec.go",
         "log.go",
-        "services.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain",
     visibility = ["//:__subpackages__"],
@@ -32,12 +31,9 @@ go_test(
         "commit_graph_test.go",
         "common_test.go",
         "exec_test.go",
-        "services_test.go",
     ],
     embed = [":gitdomain"],
     deps = [
-        "//internal/api",
-        "//lib/errors",
         "@com_github_google_go_cmp//cmp",
         "@com_github_sourcegraph_log//logtest",
         "@com_github_stretchr_testify//assert",
