+++ internal/gitserver/gitdomain/common_test.go
@@ -94,3 +94,18 @@ func TestRefGlobs(t *testing.T) {
 		})
 	}
 }
+
+func TestIsAbsoluteRevision(t *testing.T) {
+	yes := []string{"8cb03d28ad1c6a875f357c5d862237577b06e57c", "20697a062454c29d84e3f006b22eb029d730cd00"}
+	no := []string{"ref: refs/heads/appsinfra/SHEP-20-review", "master", "HEAD", "refs/heads/master", "20697a062454c29d84e3f006b22eb029d730cd0", "20697a062454c29d84e3f006b22eb029d730cd000", "  20697a062454c29d84e3f006b22eb029d730cd00  ", "20697a062454c29d84e3f006b22eb029d730cd0 "}
+	for _, s := range yes {
+		if !IsAbsoluteRevision(s) {
+			t.Errorf("%q should be an absolute revision", s)
+		}
+	}
+	for _, s := range no {
+		if IsAbsoluteRevision(s) {
+			t.Errorf("%q should not be an absolute revision", s)
+		}
+	}
+}
